Successfully  implemented displaying to user todays events #46
Since the .Day enum is saved in the database as an int in RecurId , it was initially difficult to create a LINQ query in the home/index page to convert the data to the Enum. This is because LINQ has difficulty working with enums and was not recognized by LINQ. The solution was to create a method for DateTime.Today.DayOfWeek to convert the output to an Enum "ConvertToDay" and then cast it into an int. Now saved as the variable " todayAsDayEnum" I could now use a bitwise comparison to create the correct query:  where (loggedInUserEvent.RecurID & todayAsDayEnum) == todayAsDayEnum    will take the bit value of the int and compare it with today. The end result is the program will show what are today's goals in the home page.