fixed bug that would discern exp. user only based on todays goals #56
While fixing the home page I noticed that the check for exp.  user would now only check for todays goals, so if the user had no active goals today then it would treat them as a new user.
I implemented a simple augmentation to the var currentUserEvents and called it currentUserAllGoalList. This counts all of the users goals.